Also check launch URL for fragments
Querystrings were already being checked for and passed along to the launch event args, however in some cases we might want to use fragments instead. The best use case I can think of is that they don't get sent to from the client in web requests, and they don't participate in redirects.  Anyway, I think it's a simple case that would be useful to include.